web/requirements.txt
Django
dj-database-url
dj-static
django-toolbelt
gunicorn
psycopg2
static3
django-bootstrap3
wsgiref
django-crispy-forms
pyyaml